Nest SudokuSquare class within SudokuBoard
This makes more logical sense than keeping SudokuSquare as a public
top-level class, as no other class will use SudokuSquares. Changes
to the squares will be made by calling SudokuBoard methods. This
change will likely require rewriting some existing methods.
