+++ SudokuBoard.java
@@ -1,18 +1,3 @@
-/**
- * Standard 9 x 9 sudoku board made from SudokuSquares.
- * SudokuBoard is made of SudokuSquare objects. 
- * <p>
- * Cells on the board are specified with a (row, column) in the following way:
- * (1,1) (1,2) ... (1,9)
- * (2,1) ...
- * .
- * .
- * . 
- * (9,1) (9,2) ... (9,9)
- *
- * @version 1.0
- */
-
 import java.io.BufferedReader;
 import java.io.FileReader;
 
@@ -21,7 +6,22 @@ import java.io.IOException;
 
 import java.util.Arrays;
 import java.util.ArrayList;
+import java.util.Collections;
 
+/**
+ * Standard 9 x 9 sudoku board made from SudokuSquares.
+ * SudokuBoard is made of SudokuSquare objects.
+ * <p>
+ * Cells on the board are specified with a (row, column) in the following way:
+ * (1,1) (1,2) ... (1,9)
+ * (2,1) ...
+ * .
+ * .
+ * .
+ * (9,1) (9,2) ... (9,9)
+ *
+ * @version 1.1
+ */
 public class SudokuBoard {
 
   private SudokuSquare[][] cells;
@@ -174,13 +174,125 @@ public class SudokuBoard {
     return newreg;
   }
 
+/**
+ * Basic building sudoku board building block.
+ * SudokuSquare is the class of objects that will make up a sudoku board.
+ * Each square will be able to hold a single value as well as multiple
+ * penciled values.
+ */
+  class SudokuSquare {
+    private ArrayList<Integer> pencils;
+    private int value;
+  /**
+   * Constructs an empty SudokuSquare.
+   * Value set to 0 by default.
+   */
+    public SudokuSquare() {
+      this(0);
+    }
+
+  /**
+   * Constructs a SudokuSquare with the passed value.
+   *
+   * @param value value to create the square with
+   */
+    public SudokuSquare(int value) {
+      pencils = new ArrayList<Integer>();
+      validateValue(value);
+      this.value = value;
+    }
+
+  /**
+   * Sets the square value.
+   * Value must be between 1 and 9 inclusive.
+   *
+   * @param value desired value of square
+   */
+    public void setValue(int value) {
+      validateValue(value);
+      this.value = value;
+    }
+
+  /**
+   * Returns the square value.
+   *
+   * @return the square's value
+   */
+    public int getValue() {
+      return value;
+    }
+
+  /**
+   * Adds a penciled value.
+   *
+   * @param value value to pencil in
+   */
+    public void addPencil(int value) {
+      validateValue(value);
+      pencils.add(value);
+      Collections.sort(pencils);
+    }
+
+  /**
+   * Removes a penciled value.
+   *
+   * @param value value to unpencil
+   */
+    public void removePencil(int value) {
+      validateValue(value);
+      pencils.remove(new Integer(value));
+    }
+
+  /**
+   * Returns the list of pencils at the square.
+   *
+   * @return list of pencils
+   */
+    public ArrayList<Integer> getPencils() {
+      return pencils;
+    }
+
+  /** Removes all penciled values. */
+    public void clearPencils() {
+      pencils.clear();
+    }
+
+  /**
+   * Returns (empty/full) state of square.
+   *
+   * @return true if a value other than 0, false if 0.
+   */
+    public boolean isFull() {
+      return (value != 0);
+    }
+
+  /**
+   * Allows SudokuSquares to be printable with println.
+   *
+   * @return the string representation of the square's value
+   */
+    public String toString() {
+      return Integer.toString(value);
+    }
+
+  /**
+   * Throws exceptions for a bad input.
+   *
+   * @param value value to check
+   */
+    private void validateValue(int value) {
+      if (value < 0 || value > 9) {
+        throw new IllegalArgumentException("Value must be between 0 and 9.");
+      }
+    }
+  }
 /** Method for quick testing. */
   public static void main(String[] args) {
     try {
       BufferedReader in = new BufferedReader(new FileReader(args[0]));
       String boardstring = in.readLine();
       SudokuBoard sb = new SudokuBoard(boardstring);
-      //System.out.println(sb.toPrettyString());
+      System.out.println(sb.toPrettyString());
       //System.out.println(Arrays.toString(sb.getReg(5)));
     } catch (FileNotFoundException e) {
       System.out.println("File not found");
